Initial commit for reduction of memory in diagnostics
We can now search for all unique fields within the DiagList for a
given diagnostic ID (i.e. all species for SpeciesConc, DryDepFlx,
etc.).  This will allow us to pre-compute the size of the diagnostic
arrays required in order to only allocate as much memory as is needed.

The work is in a beginning stage of development.  Still to do:
(1) Add mapping arrays for more diagnostics
(2) Make sure the algorithm works for e.g. only drydep, wetdep species
    instead of advected or all species.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>