Now implement proper mapping for diagnostic arrays
We have made modifications to several modules so that each State_Diag
array for which we would like to reduce memory will have an associated
mapping array.  This mapping array will be of the length of the number
of species (or drydep species, or advected species, or wetdep species,
etc.) and will store the indices of the corresponding State_Diag array
that we wish to archive.  All other indices will be set to -1, indicating
that those species will be not archived.

At present, this seems to work for the SpeciesConc and DryDepVel
arrays, but further testing is needed.  We have to add code in the
more of the REGISTER_DIAGFIELD_* module procedures.

Also -- it will probably be more efficient if we modify the routine
Registry_AddField so as to insert new fields into the linked list right
after the head.  This will preserve the ordering of the fields as
listed in HISTORY.rc.  This will also ensure that when we select
all species, that we step through the array in the most optimal way.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>