Add fixes to Get_Mapping routine in state_diag_mod.F90
We have updated the Get_Mapping routine so that it accepts a list of
fields (not tags).  This is necessary so that we can register the
fields by field name properly.  Also revamped the algorithm that
computes the mapping (i.e. the index in the State_Diag array
corresponding to each species number).

This update passed a geosfp_4x5_benchmark difference test.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>