Add mapping arrays for DryDepChm, DryDepMix, and DryDep diagnostics
We have added the relevant mapping arrays in State_Diag for the
DryDepChm, DryDepMix, and DryDep diagnostics.  The corresponding
diagnostic arrays in State_Diag will now be allocated only with
enough memory for species that are actually listed in HISTORY.rc.

The mapping arrays are now used in the various modules where
drydep diagnostics are updated.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>