Add mapping for AdvFlux*, WetLossConv*, and WetLossLs* diagnostics
We have added the relevant mapping arrays in State_Diag for the
following diagnostics:

  AdvFluxVert, AdvFluxZonal, and AdvFluxMerid,
  WetLossConvFrac
  WetLossConv
  WetLossLS

The corresponding diagnostic arrays in State_Diag will now be
allocated only with enough memory for species that are actually
listed in HISTORY.rc.

The mapping arrays are now used in the various modules where these
diagnostics are located.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>