Avoid out-of-bounds errors for HISTORY diagnostics in TPCORE
In tpcore_fvdas_mod.F90, we have to bracket the references to the
mapping arrays (State%Diag%Map_AdvFlux*) with IF statements.  This is
because the same block of code executes for the bpch diagnostics.
This can lead to an error if the bpch diagnostics ND24, ND25, or ND26
are turned on but the netCDF diagnostics are not.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>